requirements.txt
django
djangorestframework
djongo
sqlparse==0.2.4
graphene-django
django-graphiql
django-filter>=2
faker
sgqlc
httpie
django-cors-headers